xioami_sensor_poller/pyproject.toml
[tool.poetry]
name = "xioami_sensor_poller"
version = "0.1.0"
description = ""
authors = ["Mikael Andersen <andersen.mikael@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
mitemp_bt = { git = "https://github.com/ratcashdev/mitemp", tag = "v0.0.3" }
bluepy = { git = "https://github.com/IanHarvey/bluepy", tag = "v/1.3.0" }
paho-mqtt = "1.5.1" 

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tempPoller = "tempPoller:main"